Req OLTSWITCH-1435 (merge from trunk r8162-8163, 8172-8173) Integrate NBTools Logger in the 3.17 version (TC16SXG branch):  - Corrections were made to all inconsistent LOGs in order to avoid errors when using the prototype    'extern log_ret_t logger_print(...) __attribute__((format(printf, 7, 8)));'  - Corrections applied to switchdrvr's generic files.
svn path=/branches/3.17_tc16sxg/; revision=8175

+++ src/fastpath/andl/hapi/esw/usl/policy/l7_usl_policy_db.c
@@ -483,14 +483,14 @@ void usl_policy_db_memory_info()
   printf("BROAD_MAX_RULES_PER_POLICY         == %d\n", BROAD_MAX_RULES_PER_POLICY);
   printf("BROAD_MAX_SYSTEM_POLICIES          == %d\n", BROAD_MAX_SYSTEM_POLICIES);
   printf("sizeof(policy_map_table_t)         == %d\n", policy_map_table_t_size());
-  printf("sizeof(BROAD_POLICY_ENTRY_t)       == %d\n", (unsigned int) sizeof(BROAD_POLICY_ENTRY_t));
-  printf("sizeof(BROAD_POLICY_RULE_ENTRY_t)  == %d\n", (unsigned int) sizeof(BROAD_POLICY_RULE_ENTRY_t));
-  printf("sizeof(BROAD_FIELD_ENTRY_t)        == %d\n", (unsigned int) sizeof(BROAD_FIELD_ENTRY_t));
-  printf("sizeof(BROAD_ACTION_ENTRY_t)       == %d\n", (unsigned int) sizeof(BROAD_ACTION_ENTRY_t));
+  printf("sizeof(BROAD_POLICY_ENTRY_t)       == %zu\n", sizeof(BROAD_POLICY_ENTRY_t));
+  printf("sizeof(BROAD_POLICY_RULE_ENTRY_t)  == %zu\n", sizeof(BROAD_POLICY_RULE_ENTRY_t));
+  printf("sizeof(BROAD_FIELD_ENTRY_t)        == %zu\n", sizeof(BROAD_FIELD_ENTRY_t));
+  printf("sizeof(BROAD_ACTION_ENTRY_t)       == %zu\n", sizeof(BROAD_ACTION_ENTRY_t));
   printf("\n\n");
 
-  printf("policy_map_table[]          == %d\n", BROAD_MAX_BCM_UNITS_PER_CPU * policy_map_table_t_size() * BROAD_MAX_POLICIES_PER_BCM_UNIT);
-  printf("pUslOperPolicyDB[]          == %d\n", (unsigned int) sizeof(BROAD_USL_POLICY_t) * BROAD_MAX_POLICIES);
+  printf("policy_map_table[]          == %u\n", BROAD_MAX_BCM_UNITS_PER_CPU * policy_map_table_t_size() * BROAD_MAX_POLICIES_PER_BCM_UNIT);
+  printf("pUslOperPolicyDB[]          == %zu\n", sizeof(BROAD_USL_POLICY_t) * BROAD_MAX_POLICIES);
 }
 
 /*********************************************************************
