+++ src/fastpath/src/application/qos/diffserv/base/diffserv_callback.c
@@ -507,7 +507,7 @@ static L7_RC_t dsCommonCallbackTableShow(diffServCallbackTableId_t tableId,
       }
       DIFFSERV_PRT(msgLvlReqd,
               " [%2u] funcPtr=0x%llx comp=%-*s descr=%s\n",
-              i, PTR_TO_UINT64(pCbEntry->funcPtr), sizeof(compIdStr), compIdStr, pCbEntry->displayStr);
+              i, PTR_TO_UINT64(pCbEntry->funcPtr), (int) sizeof(compIdStr), compIdStr, pCbEntry->displayStr);
     }
   } /* endfor */
